#1 Mockable dialogs, test create entry dialog

+++ lib/widget/dialog/delete_entry_dialog.dart
@@ -3,8 +3,8 @@ import 'package:free_authenticator/model/api/entry.dart';
 import 'package:free_authenticator/model/api/entry_type.dart';
 import 'package:free_authenticator/widget/dependencies.dart';
 
-class DeleteEntry extends StatefulWidget {
-  DeleteEntry({
+class DeleteEntryDialog extends StatefulWidget {
+  DeleteEntryDialog({
     Key key,
     @required this.entry,
     this.onDelete,
@@ -17,7 +17,7 @@ class DeleteEntry extends StatefulWidget {
   _DeleteEntry createState() => _DeleteEntry();
 }
 
-class _DeleteEntry extends State<DeleteEntry> {
+class _DeleteEntry extends State<DeleteEntryDialog> {
   Widget _buildConfirm() {
     return AlertDialog(
       title: Text('Deletion'),
@@ -30,7 +30,7 @@ class _DeleteEntry extends State<DeleteEntry> {
           onPressed: () async {
             await Dependencies.of(context).store.deleteEntry(this.widget.entry.id);
             Navigator.of(context).pop();
-            this.widget.onDelete(this.widget.entry.id);
+            if (widget.onDelete != null) this.widget.onDelete(this.widget.entry.id);
           },
         ),
         new FlatButton(
